Finish UI for rating images

+++ sql/007_ReviewPictures.sql
@@ -23,10 +23,10 @@ CREATE TABLE review_pictures_comment
 
 CREATE TABLE review_pictures_rating
 (
-    person   BIGINT       NOT NULL,
-    folder   CHAR(36)     NOT NULL,
-    filename VARCHAR(128) NOT NULL,
-    rating   INTEGER      NOT NULL,
+    person   BIGINT        NOT NULL,
+    folder   CHAR(36)      NOT NULL,
+    filename VARCHAR(128)  NOT NULL,
+    rating   DECIMAL(2, 1) NOT NULL,
     PRIMARY KEY (person, folder, filename),
     FOREIGN KEY (person, folder) REFERENCES review_pictures_access (person, folder)
         ON UPDATE CASCADE
